% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_survival.R
\name{analyse_survival}
\alias{analyse_survival}
\title{analyse_survival}
\usage{
analyse_survival(RD, name = "no name", xtreat = NULL, c_range = c(5, 100))
}
\arguments{
\item{RD}{data.frame or matrix containing a table of experiment data}

\item{name}{experiment name (e.g. name of cell line)}

\item{xtreat}{treatment dose of the colonies counted in the corresponding
columns of RD}

\item{c_range}{number or vector of numbers of colonies counted for which
the survival fraction is to be calculated (default = c(5,100))}
}
\value{
list object containing several experiments and treatments organized
  for convenient plotting
}
\description{
clonogenic survival data from the colony formation assay is
  processed with respect to cell cooperativity and robust calculation of
  survival fractions
}
\examples{
seeded <- rep(10^(seq(1,5,0.5)),each = 3)
counted1 <- 0.4 * seeded^1.1 * rnorm(n = length(seeded),1,0.05)
counted2 <- 0.2 * seeded^1.125 * rnorm(n = length(seeded),1,0.05)
counted3 <- 0.05 * seeded^1.25 * rnorm(n = length(seeded),1,0.05)
df.1 <- data.frame("seeded" = seeded,
              "counted1" = counted1,
              "counted2" = counted2,
              "counted3" = counted3)
seeded <- rep(10^(seq(1,5,0.5)),each = 3)
counted1 <- 0.5 * seeded^1.01 * rnorm(n = length(seeded),1,0.05)
counted2 <- 0.4 * seeded^1.0125 * rnorm(n = length(seeded),1,0.05)
counted3 <- 0.2 * seeded^1.025 * rnorm(n = length(seeded),1,0.05)
df.2 <- data.frame("seeded" = seeded,
              "counted1" = counted1,
              "counted2" = counted2,
              "counted3" = counted3)
SF <- vector("list",2)
SF[[1]] <- analyse_survival(RD = df.1,
                            name = "cell line a",
                            xtreat = c(0,1,4),
                            c_range = c(5,100))
SF[[2]] <- analyse_survival(RD = df.2,
                            name = "cell line b",
                            xtreat = c(0,1,4),
                            c_range = c(5,100))
}
