% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CFAdata.R
\docType{data}
\name{CFAdata}
\alias{CFAdata}
\title{Colony Formation Assay data on cell cooperativity}
\format{
\code{data.frame}
}
\usage{
data(CFAdata)
}
\description{
Clonogenic survival data from seven cell lines T47D, MDA-MB231, A549,
HCC1806, SKBR3, SKLU1 and BT20 as presented in Brix et al. (2020).
}
\examples{
data(CFAdata)
head(CFAdata)
cll <- levels(CFAdata$cell.line)
}
\references{
Brix et al. (2020) Radiation Oncology, "The clonogenic assay:
robustness of plating efficiency-based analysis is strongly compromised by
cellular cooperation"
}
\keyword{dataset}
