% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_SF.R
\name{plot_SF}
\alias{plot_SF}
\title{plot_SF}
\usage{
plot_SF(SF)
}
\arguments{
\item{SF}{list build from objects returned by analyse_survival}
}
\value{
none
}
\description{
plot cellular cooperativity and clonogenic survival data from
colony formation assay data
}
\examples{
seeded <- rep(10^(seq(1,5,0.5)),each = 3)
counted1 <- 0.4 * seeded^1.1 * rnorm(n = length(seeded),1,0.05)
counted2 <- 0.2 * seeded^1.125 * rnorm(n = length(seeded),1,0.05)
counted3 <- 0.05 * seeded^1.25 * rnorm(n = length(seeded),1,0.05)
df.1 <- data.frame("seeded" = seeded,
              "counted1" = counted1,
              "counted2" = counted2,
              "counted3" = counted3)
seeded <- rep(10^(seq(1,5,0.5)),each = 3)
counted1 <- 0.5 * seeded^1.01 * rnorm(n = length(seeded),1,0.05)
counted2 <- 0.4 * seeded^1.0125 * rnorm(n = length(seeded),1,0.05)
counted3 <- 0.2 * seeded^1.025 * rnorm(n = length(seeded),1,0.05)
df.2 <- data.frame("seeded" = seeded,
              "counted1" = counted1,
              "counted2" = counted2,
              "counted3" = counted3)
SF <- vector("list",2)
SF[[1]] <- analyse_survival(RD = df.1,name = "cell line a",xtreat = c(0,1,4))
SF[[2]] <- analyse_survival(RD = df.2,name = "cell line b",xtreat = c(0,1,4))
plot_SF(SF)
}
