% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CFAdata.R
\docType{data}
\name{CFAdata}
\alias{CFAdata}
\title{Colony Formation Assay data on cell cooperativity}
\format{
\code{data.frame}
}
\usage{
data(CFAdata)
}
\description{
Clonogenic Survival Data from seven cell lines
}
\examples{
data(CFAdata)
head(CFAdata)
cll <- levels(CFAdata$cell.line)
}
\references{
Brix et al. (2020) Radiation Oncology ...
}
\keyword{dataset}
