% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwr_reg.R
\name{pwr_reg}
\alias{pwr_reg}
\title{pwr_reg}
\usage{
pwr_reg(x)
}
\arguments{
\item{x}{data.frame or matrix of dimension n x 2:
first column contains number of cells seeded,
second to n-th column contains number of colonies counted}
}
\value{
\code{summary.lm} object as returned by \code{\link{summary}}
}
\description{
\code{pwr_reg} calculates a power regression for cell
  cooperativity experiments from Colony Formation Assay (CFA) data
}
\examples{
df.xmpl <- data.frame("seeded" = 10^(seq(1,5,0.5)),
              "counted" = 0.4*10^(seq(1,5,0.5)+rnorm(n = 9,0,0.1))^1.25)
pwr_reg(x = df.xmpl)
}
